Move `react-docs` into `website` and delete it as a standalone package
