+++ .github/autolabeler.yml
@@ -22,7 +22,6 @@ contracts: ['contracts']
 @0x/tslint-config: ['packages/tslint-config']
 @0x/asset-buyer: ['packages/asset-buyer']
 @0x/order-watcher: ['packages/order-watcher']
-@0x/react-docs: ['packages/react-docs']
 @0x/order-utils: ['packages/order-utils']
 @0x/react-shared: ['packages/react-shared']
 @0x/assert: ['packages/assert']
