--- packages/react-docs/.npmignore
@@ -1,9 +1,0 @@
-# Blacklist all files
-.*
-*
-# Whitelist lib
-!lib/**/*
-# Blacklist tests and publish scripts
-/lib/test/*
-/lib/monorepo_scripts/
-# Package specific ignore
