--- packages/react-docs/CHANGELOG.md
@@ -1,240 +1,0 @@
-<!--
-changelogUtils.file is auto-generated using the monorepo-scripts package. Don't edit directly.
-Edit the package's CHANGELOG.json file only.
--->
-
-CHANGELOG
-
-## v2.0.15 - _July 24, 2019_
-
-    * Dependencies updated
-
-## v2.0.14 - _July 13, 2019_
-
-    * Dependencies updated
-
-## v2.0.13 - _July 13, 2019_
-
-    * Dependencies updated
-
-## v2.0.12 - _May 24, 2019_
-
-    * Dependencies updated
-
-## v2.0.11 - _May 10, 2019_
-
-    * Dependencies updated
-
-## v2.0.10 - _April 11, 2019_
-
-    * Dependencies updated
-
-## v2.0.9 - _March 21, 2019_
-
-    * Dependencies updated
-
-## v2.0.8 - _March 20, 2019_
-
-    * Dependencies updated
-
-## v2.0.7 - _March 1, 2019_
-
-    * Dependencies updated
-
-## v2.0.6 - _February 26, 2019_
-
-    * Dependencies updated
-
-## v2.0.5 - _February 25, 2019_
-
-    * Dependencies updated
-
-## v2.0.4 - _February 9, 2019_
-
-    * Dependencies updated
-
-## v2.0.3 - _February 7, 2019_
-
-    * Dependencies updated
-
-## v2.0.2 - _February 7, 2019_
-
-    * Dependencies updated
-
-## v2.0.1 - _February 6, 2019_
-
-    * Dependencies updated
-
-## v2.0.0 - _February 5, 2019_
-
-    * Upgrade the bignumber.js to v8.0.2 (#1517)
-
-## v1.0.25 - _January 15, 2019_
-
-    * Dependencies updated
-
-## v1.0.24 - _January 11, 2019_
-
-    * Dependencies updated
-
-## v1.0.23 - _January 9, 2019_
-
-    * Dependencies updated
-
-## v1.0.22 - _December 13, 2018_
-
-    * Dependencies updated
-
-## v1.0.21 - _December 11, 2018_
-
-    * Dependencies updated
-
-## v1.0.20 - _November 28, 2018_
-
-    * Dependencies updated
-
-## v1.0.19 - _November 21, 2018_
-
-    * Dependencies updated
-
-## v1.0.18 - _November 14, 2018_
-
-    * Dependencies updated
-
-## v1.0.17 - _November 13, 2018_
-
-    * Dependencies updated
-
-## v1.0.16 - _November 12, 2018_
-
-    * Dependencies updated
-
-## v1.0.15 - _November 9, 2018_
-
-    * Dependencies updated
-
-## v1.0.14 - _October 18, 2018_
-
-    * Dependencies updated
-
-## v1.0.13 - _October 4, 2018_
-
-    * Dependencies updated
-
-## v1.0.12 - _September 28, 2018_
-
-    * Dependencies updated
-
-## v1.0.11 - _September 25, 2018_
-
-    * Dependencies updated
-
-## v1.0.10 - _September 25, 2018_
-
-    * Dependencies updated
-
-## v1.0.9 - _September 21, 2018_
-
-    * Dependencies updated
-
-## v1.0.8 - _September 5, 2018_
-
-    * Dependencies updated
-
-## v1.0.7 - _August 27, 2018_
-
-    * Dependencies updated
-
-## v1.0.6 - _August 24, 2018_
-
-    * Dependencies updated
-
-## v1.0.5 - _August 14, 2018_
-
-    * Dependencies updated
-
-## v1.0.4 - _July 26, 2018_
-
-    * Dependencies updated
-
-## v1.0.3 - _July 26, 2018_
-
-    * Dependencies updated
-
-## v1.0.2 - _July 26, 2018_
-
-    * Dependencies updated
-
-## v1.0.1 - _July 23, 2018_
-
-    * Dependencies updated
-
-## v1.0.0 - _July 19, 2018_
-
-    * Dependencies updated
-
-## v0.0.17 - _July 18, 2018_
-
-    * Nest MD files under versions so that you can update them for future versions (#844)
-
-## v0.0.16 - _July 18, 2018_
-
-    * Dependencies updated
-
-## v0.0.15 - _July 9, 2018_
-
-    * Dependencies updated
-
-## v0.0.14 - _June 19, 2018_
-
-    * Dependencies updated
-
-## v0.0.13 - _May 31, 2018_
-
-    * Incorrect publish that was unpublished
-
-## v0.0.12 - _May 22, 2018_
-
-    * Dependencies updated
-
-## v0.0.11 - _May 4, 2018_
-
-    * Dependencies updated
-
-## v0.0.10 - _May 4, 2018_
-
-    * Dependencies updated
-
-## v0.0.9 - _May 4, 2018_
-
-    * Dependencies updated
-
-## v0.0.8 - _April 18, 2018_
-
-    * Added support for rendering default param values (#519)
-    * Added support for rendering nested function types within interface types (#519)
-    * Improve type comment rendering (#535)
-
-## v0.0.7 - _April 11, 2018_
-
-    * Dependencies updated
-
-## v0.0.6 - _April 2, 2018_
-
-    * Dependencies updated
-
-## v0.0.5 - _April 2, 2018_
-
-    * Handle `reflection` type rendering so that anonymous function type declarations render properly (#465)
-    * Rename `MethodSignature` to `Signature` and change it's props so that it can be used to render method and function signatures. (#465)
-    * Rename `MethodBlock` to `SignatureBlock` since it is not used to render method and function signature blocks. (#465)
-    * Add support for documenting exported functions. (#465)
-
-## v0.0.3 - _March 17, 2018_
-
-    * Move TS typings from devDependencies to dependencies since they are needed by the package user.
-
-## v0.0.2 - _March 17, 2018_
-
-    * Move example out into a separate sub-package
-    * Consolidate all `console.log` calls into `logUtils` in the `@0xproject/utils` package (#452)
