--- packages/react-docs/package.json
@@ -1,62 +1,0 @@
-{
-    "name": "@0x/react-docs",
-    "version": "2.0.15",
-    "engines": {
-        "node": ">=6.12"
-    },
-    "description": "React documentation component for rendering TypeDoc & sol-doc generated JSON",
-    "main": "lib/index.js",
-    "types": "lib/index.d.ts",
-    "scripts": {
-        "lint": "tslint --format stylish --project .",
-        "fix": "tslint --fix --format stylish --project .",
-        "build": "tsc -b",
-        "build:ci": "yarn build",
-        "clean": "shx rm -rf lib"
-    },
-    "author": "Fabio Berger",
-    "license": "Apache-2.0",
-    "bugs": {
-        "url": "https://github.com/0xProject/0x-monorepo/issues"
-    },
-    "homepage": "https://github.com/0xProject/0x-monorepo/packages/react-docs/README.md",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/0xProject/0x-monorepo.git"
-    },
-    "devDependencies": {
-        "@0x/dev-utils": "^2.2.5",
-        "@0x/tslint-config": "^3.0.1",
-        "@types/compare-versions": "^3.0.0",
-        "@types/styled-components": "4.1.1",
-        "make-promises-safe": "^1.1.0",
-        "shx": "^0.2.2",
-        "tslint": "5.11.0",
-        "typescript": "3.0.1"
-    },
-    "dependencies": {
-        "@0x/react-shared": "^2.0.15",
-        "@0x/types": "^2.4.1",
-        "@0x/utils": "^4.4.1",
-        "@types/lodash": "4.14.104",
-        "@types/material-ui": "^0.20.0",
-        "@types/node": "*",
-        "@types/react": "*",
-        "@types/react-dom": "*",
-        "@types/react-scroll": "1.5.3",
-        "basscss": "^8.0.3",
-        "compare-versions": "^3.0.1",
-        "lodash": "^4.17.11",
-        "material-ui": "^0.20.0",
-        "react": "^16.5.2",
-        "react-dom": "^16.5.2",
-        "react-markdown": "^3.2.2",
-        "react-scroll": "0xproject/react-scroll#pr-330-and-replace-state",
-        "react-tooltip": "^3.2.7",
-        "semver": "5.5.0",
-        "styled-components": "^4.1.1"
-    },
-    "publishConfig": {
-        "access": "public"
-    }
-}
