--- packages/react-docs/src/globals.d.ts
@@ -1,6 +1,0 @@
-declare module '*.json' {
-    const json: any;
-    /* tslint:disable */
-    export default json;
-    /* tslint:enable */
-}
