--- packages/react-docs/src/index.ts
@@ -1,23 +1,0 @@
-export { DocAgnosticFormat, GeneratedDocJson } from '@0x/types';
-
-// Exported to give users of this library added flexibility if they want to build
-// a docs page from scratch using the individual components.
-export { Badge } from './components/badge';
-export { Comment } from './components/comment';
-export { CustomEnum } from './components/custom_enum';
-export { DocReference } from './components/doc_reference';
-export { Enum } from './components/enum';
-export { EventDefinition } from './components/event_definition';
-export { Interface } from './components/interface';
-export { SignatureBlock } from './components/signature_block';
-export { Signature } from './components/signature';
-export { SourceLink } from './components/source_link';
-export { TypeDefinition } from './components/type_definition';
-export { Type } from './components/type';
-export { TypeDocUtils } from './utils/typedoc_utils';
-
-export { DocsInfo } from './docs_info';
-
-export { DocsInfoConfig, DocsMenu, SupportedDocJson } from './types';
-
-export { constants } from './utils/constants';
