--- packages/react-docs/src/utils/constants.ts
@@ -1,10 +1,0 @@
-import { SupportedDocJson } from '../types';
-
-export const constants = {
-    TYPES_SECTION_NAME: 'types',
-    EXTERNAL_EXPORTS_SECTION_NAME: 'external exports',
-    TYPE_TO_SYNTAX: {
-        [SupportedDocJson.SolDoc]: 'solidity',
-        [SupportedDocJson.TypeDoc]: 'typescript',
-    } as { [supportedDocType: string]: string },
-};
