--- packages/react-docs/tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "extends": "../../tsconfig",
-    "compilerOptions": {
-        "outDir": "lib",
-        "rootDir": "src",
-        "jsx": "react",
-        "baseUrl": ".",
-        "strictNullChecks": false,
-        "paths": {
-            "*": ["node_modules/@types/*", "*"]
-        }
-    },
-    "include": ["./src/**/*"]
-}
