--- packages/react-docs/tslint.json
@@ -1,9 +1,0 @@
-{
-    "extends": ["@0x/tslint-config"],
-    "rules": {
-        "no-object-literal-type-assertion": false,
-        "completed-docs": false,
-        "prefer-function-over-method": false,
-        "custom-no-magic-numbers": false
-    }
-}
