+++ packages/website/ts/components/documentation/reference/interface.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 
 import { CustomType, TypeDefinitionByName } from '@0x/types';
 
-import { DocsInfo } from '../docs_info';
+import { DocsInfo } from '../../../utils/docs_info';
 
 import { Signature } from './signature';
 import { Type } from './type';
