+++ packages/website/ts/components/documentation/reference/signature.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 
 import { Parameter, Type as TypeDef, TypeDefinitionByName, TypeParameter } from '@0x/types';
 
-import { DocsInfo } from '../docs_info';
+import { DocsInfo } from '../../../utils/docs_info';
 
 import { Type } from './type';
