+++ packages/website/ts/containers/contract_wrappers_documentation.ts
@@ -1,9 +1,9 @@
-import { DocsInfoConfig, SupportedDocJson } from '@0x/react-docs';
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { DocPage as DocPageComponent, DocPageProps } from 'ts/pages/documentation/doc_page';
 import { DocPackages } from 'ts/types';
 
+import { DocsInfoConfig, SupportedDocJson } from '../types';
 import { getMapStateToProps, mapDispatchToProps } from '../utils/documentation_container';
 
 /* tslint:disable:no-var-requires */
