+++ packages/website/ts/containers/smart_contracts_documentation.ts
@@ -1,10 +1,10 @@
-import { DocsInfoConfig, SupportedDocJson } from '@0x/react-docs';
 import { Networks } from '@0x/react-shared';
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { DocPage as DocPageComponent, DocPageProps } from 'ts/pages/documentation/doc_page';
 import { DocPackages, SmartContractDocSections as Sections } from 'ts/types';
 
+import { DocsInfoConfig, SupportedDocJson } from '../types';
 import { getMapStateToProps, mapDispatchToProps } from '../utils/documentation_container';
 
 /* tslint:disable:no-var-requires */
